[v0.1.1] Clean up folder structure & resolve bugs
* All JavaScript-based files are now ESLint-compliant.
* Users context menu now injects into the
'FluxContainer(Avatar)'. This resolves the bug where avatars of direct
messages and 'My Account' would open the user accounts context menu.
* Around line 45 of the base file, 'index.js', now declares a default value. No
more crashes from here on.
* Additionally, somewhere near line 89 of the settings file, 'Settings.jsx', uses the
`defaultValue` attribute rather than `value`.
