+++ manifest.json
@@ -1,6 +1,6 @@
 {
   "name": "Mu",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "description": "Open separate instances of Discord with different users.",
   "author": "halcyxn",
   "license": "LGPLv3"
